+++ src/Extensions/PostalCodeFor.php
@@ -2,7 +2,7 @@
 
 namespace Axlon\PostalCodeValidation\Extensions;
 
-use Axlon\PostalCodeValidation\PatternMatcher;
+use Axlon\PostalCodeValidation\PostalCodeValidator;
 use Axlon\PostalCodeValidation\Support\PostalCodeExamples;
 use Illuminate\Support\Arr;
 use Illuminate\Validation\Validator;
@@ -20,21 +20,21 @@ class PostalCodeFor
     /**
      * The pattern matcher.
      *
-     * @var \Axlon\PostalCodeValidation\PatternMatcher
+     * @var \Axlon\PostalCodeValidation\PostalCodeValidator
      */
-    protected $matcher;
+    protected $validator;
 
     /**
      * Create a new PostalCodeFor validator extension.
      *
-     * @param \Axlon\PostalCodeValidation\PatternMatcher $matcher
+     * @param \Axlon\PostalCodeValidation\PostalCodeValidator $validator
      * @param \Axlon\PostalCodeValidation\Support\PostalCodeExamples $examples
      * @return void
      */
-    public function __construct(PatternMatcher $matcher, PostalCodeExamples $examples)
+    public function __construct(PostalCodeValidator $validator, PostalCodeExamples $examples)
     {
         $this->examples = $examples;
-        $this->matcher = $matcher;
+        $this->validator = $validator;
     }
 
     /**
@@ -106,7 +106,7 @@ class PostalCodeFor
                 continue;
             }
 
-            if ($this->matcher->passes($parameter, $value)) {
+            if ($this->validator->passes($parameter, $value)) {
                 return true;
             }
         }