+++ src/PostalCodeValidator.php
@@ -2,7 +2,7 @@
 
 namespace Axlon\PostalCodeValidation;
 
-class PatternMatcher
+class PostalCodeValidator
 {
     /**
      * The matching patterns.
@@ -53,7 +53,8 @@ class PatternMatcher
     {
         if (is_array($countryCode)) {
             $this->patternOverrides = array_merge(
-                $this->patternOverrides, array_change_key_case($countryCode, CASE_UPPER)
+                $this->patternOverrides,
+                array_change_key_case($countryCode, CASE_UPPER)
             );
         } else {
             $this->patternOverrides[strtoupper($countryCode)] = $pattern;